+++ src/Colors/Oklab/Decoders/StringColorDecoder.php
@@ -45,7 +45,7 @@ class StringColorDecoder extends AbstractDecoder implements DecoderInterface
     public function decode(mixed $input): ColorInterface
     {
         if (preg_match(self::OKLAB_PATTERN, $input, $matches) != 1) {
-            throw new InvalidArgumentException('Invalid oklab() color notation');
+            throw new InvalidArgumentException('Invalid oklab() color syntax');
         }
 
         return new Color(...[