+++ src/Colors/Oklch/Decoders/StringColorDecoder.php
@@ -46,7 +46,7 @@ class StringColorDecoder extends AbstractDecoder implements DecoderInterface
     public function decode(mixed $input): ColorInterface
     {
         if (preg_match(self::PATTERN, $input, $matches) != 1) {
-            throw new InvalidArgumentException('Invalid oklch() color notation');
+            throw new InvalidArgumentException('Invalid oklch() color syntax');
         }
 
         return new Color(...[