+++ changelog.md
@@ -12,7 +12,7 @@
 - Support for OkLab color space
 - Improved structure and hiarchy for exceptions
 - Improved error and exception messages
-- Function color format now supports notation without comma
+- Function color format now supports syntax without comma
 
 ## API Changes